BCD-Services/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
idna==3.10
pip==23.2.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0